+++ spinetoolbox/widgets/graph_view_graphics_items.py
@@ -105,6 +105,10 @@ class EntityItem(QGraphicsPixmapItem):
     def boundingRect(self):
         return super().boundingRect() | self.childrenBoundingRect()
 
+    def moveBy(self, x, y):
+        super().moveBy(x, y)
+        self.update_arcs_line()
+
     def _init_bg(self):
         self._bg = QGraphicsRectItem(self.boundingRect(), self)
         self._bg.setPen(Qt.NoPen)
@@ -245,7 +249,6 @@ class EntityItem(QGraphicsPixmapItem):
         for item in self.scene().selectedItems():
             if isinstance(item, (EntityItem)):
                 item.moveBy(move_by.x(), move_by.y())
-                item.move_arc_items(move_by)
         self._merge_target = self._find_merge_target()
         for view in self.scene().views():
             self._views_cursor.setdefault(view, view.viewport().cursor())
@@ -260,13 +263,10 @@ class EntityItem(QGraphicsPixmapItem):
             else:
                 view.viewport().setCursor(Qt.ForbiddenCursor)
 
-    def move_arc_items(self, pos_diff):
-        """Moves arc items.
-
-        Args:
-            pos_diff (QPoint)
-        """
-        raise NotImplementedError()
+    def update_arcs_line(self):
+        """Moves arc items."""
+        for item in self.arc_items:
+            item.update_line()
 
     def mouseReleaseEvent(self, event):
         """Merges the item into the registered target if any. Bounces it if not possible.
@@ -279,23 +279,21 @@ class EntityItem(QGraphicsPixmapItem):
         if self._merge_target:
             if self.merge_into_target():
                 return
-            self._bounce_back(self.pos())
+            self._bounce_back()
         if self._moved_on_scene:
             self._moved_on_scene = False
             scene = self.scene()
             scene.shrink_if_needed()
             scene.item_move_finished.emit(self)
 
-    def _bounce_back(self, current_pos):
+    def _bounce_back(self):
         """Bounces the item back from given position to its original position.
-
-        Args:
-            current_pos (QPoint)
         """
         if self._press_pos is None:
             return
-        self.move_arc_items(self._press_pos - current_pos)
-        self.setPos(self._press_pos)
+        delta = self._press_pos - self.pos()
+        self.moveBy(delta.x(), delta.y())
+        self.update_arcs_line()
 
     def itemChange(self, change, value):
         """
@@ -401,15 +399,6 @@ class RelationshipItem(EntityItem):
         self.become_whole()
         return True
 
-    def move_arc_items(self, pos_diff):
-        """Moves arc items.
-
-        Args:
-            pos_diff (QPoint)
-        """
-        for item in self.arc_items:
-            item.move_rel_item_by(pos_diff)
-
     def _make_wip_tool_tip(self):
         return """
             <html>This is a work-in-progress <b>{0}</b> relationship.</html>
@@ -425,6 +414,10 @@ class RelationshipItem(EntityItem):
 
     def _show_item_context_menu_in_parent(self, pos):
         self._graph_view_form.show_relationship_item_context_menu(pos)
+
+    def follow_object_by(self, x, y):
+        factor = 1.0 / len(self.arc_items)
+        self.moveBy(factor * x, factor * y)
 
 
 class ObjectItem(EntityItem):
@@ -509,15 +502,6 @@ class ObjectItem(EntityItem):
             self._graph_view_form.update_object(self.entity_id, text)
         self.refresh_name()
 
-    def move_arc_items(self, pos_diff):
-        """Moves arc items.
-
-        Args:
-            pos_diff (QPoint)
-        """
-        for item in self.arc_items:
-            item.move_obj_item_by(pos_diff)
-
     def keyPressEvent(self, event):
         """Starts editing the name if F2 is pressed.
 
@@ -579,12 +563,23 @@ class ObjectItem(EntityItem):
                 arc_item.obj_item = self
             return False
         self._merge_target.arc_items.extend(self.arc_items)
-        self.move_arc_items(self._merge_target.pos() - self.pos())
+        self.update_arcs_line()
         self.scene().removeItem(self)
         return True
 
     def _show_item_context_menu_in_parent(self, pos):
         self._graph_view_form.show_object_item_context_menu(pos, self)
+
+    def moveBy(self, x, y):
+        """Moves arc items."""
+        super().moveBy(x, y)
+        rel_icon_folows = self._graph_view_form.qsettings.value(
+            "appSettings/relationshipIconFollows", defaultValue="true"
+        )
+        if rel_icon_folows == "false":
+            return
+        for arc_item in self.arc_items:
+            arc_item.rel_item.follow_object_by(x, y)
 
 
 class EntityLabelItem(QGraphicsTextItem):
@@ -695,11 +690,7 @@ class ArcItem(QGraphicsLineItem):
         self.obj_item = obj_item
         self._width = float(width)
         self.is_wip = is_wip
-        src_x = rel_item.x()
-        src_y = rel_item.y()
-        dst_x = obj_item.x()
-        dst_y = obj_item.y()
-        self.setLine(src_x, src_y, dst_x, dst_y)
+        self.update_line()
         self._pen = QPen()
         self._pen.setWidth(self._width)
         color = QGuiApplication.palette().color(QPalette.Normal, QPalette.WindowText)
@@ -714,6 +705,13 @@ class ArcItem(QGraphicsLineItem):
         if self.is_wip:
             self.become_wip()
         self.setCursor(Qt.ArrowCursor)
+
+    def update_line(self):
+        src_x = self.rel_item.x()
+        src_y = self.rel_item.y()
+        dst_x = self.obj_item.x()
+        dst_y = self.obj_item.y()
+        self.setLine(src_x, src_y, dst_x, dst_y)
 
     def mousePressEvent(self, event):
         """Accepts the event so it's not propagated."""
@@ -733,26 +731,6 @@ class ArcItem(QGraphicsLineItem):
         self.is_wip = False
         self._pen.setStyle(Qt.SolidLine)
         self.setPen(self._pen)
-
-    def move_rel_item_by(self, pos_diff):
-        """Moves source point.
-
-        Args:
-            pos_diff (QPoint)
-        """
-        line = self.line()
-        line.setP1(line.p1() + pos_diff)
-        self.setLine(line)
-
-    def move_obj_item_by(self, pos_diff):
-        """Moves destination point.
-
-        Args:
-            pos_diff (QPoint)
-        """
-        line = self.line()
-        line.setP2(line.p2() + pos_diff)
-        self.setLine(line)
 
     def adjust_to_zoom(self, transform):
         """Adjusts the item's geometry so it stays the same size after performing a zoom.